requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
imbalanced-learn==0.12.3
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
seaborn==0.13.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rich==13.7.1
scikeras==0.13.0
scikit-learn==1.5.0
xgboost==2.0.3
scipy==1.13.1
six==1.16.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
ccxt
celery==5.3.6
flask
redis
gymnasium
# pandas-ta  # Не совместимо с Python 3.10, используем альтернативы
sqlalchemy
wandb
psycopg2-binary
binance_historical_data
alembic
PySocks
docker
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
numpy
python-dotenv
setuptools>=70
tianshou==0.5.1
tinkoff-investments>=0.2.0